Refactor generation service exports
