+++ app/frontend/src/composables/generation/createGenerationOrchestrator.ts
@@ -5,8 +5,10 @@ import {
   useGenerationTransport,
   type GenerationNotificationAdapter,
 } from './useGenerationTransport';
-import type { GenerationQueueClient } from '@/services/generation/queueClient';
-import type { GenerationWebSocketManager } from '@/services/generation/websocketManager';
+import type {
+  GenerationQueueClient,
+  GenerationWebSocketManager,
+} from '@/services';
 import { acquireSystemStatusController, DEFAULT_HISTORY_LIMIT } from '@/stores/generation';
 import type {
   GenerationConnectionStore,
