+++ app/frontend/src/composables/generation/useGenerationOrchestratorManager.ts
@@ -6,8 +6,10 @@ import {
   type GenerationOrchestrator,
 } from './createGenerationOrchestrator';
 import type { GenerationNotificationAdapter } from './useGenerationTransport';
-import type { GenerationQueueClient } from '@/services/generation/queueClient';
-import type { GenerationWebSocketManager } from '@/services/generation/websocketManager';
+import type {
+  GenerationQueueClient,
+  GenerationWebSocketManager,
+} from '@/services';
 import {
   useGenerationConnectionStore,
   useGenerationFormStore,
