+++ app/frontend/src/composables/generation/useGenerationUpdates.ts
@@ -5,8 +5,10 @@ import {
   useGenerationOrchestratorManager,
   type GenerationOrchestratorBinding,
 } from './useGenerationOrchestratorManager';
-import type { GenerationQueueClient } from '@/services/generation/queueClient';
-import type { GenerationWebSocketManager } from '@/services/generation/websocketManager';
+import type {
+  GenerationQueueClient,
+  GenerationWebSocketManager,
+} from '@/services';
 import type {
   GenerationJob,
   GenerationRequestPayload,
