+++ app/frontend/src/services/generation/index.ts
@@ -1,5 +1,36 @@
-export * from './updates';
-export * from './queueClient';
-export * from './websocketManager';
-export * from './generationService';
-export * from './validation';
+export { DEFAULT_POLL_INTERVAL, extractGenerationErrorMessage } from './updates';
+
+export {
+  createGenerationQueueClient,
+  type GenerationQueueClient,
+  type GenerationQueueClientOptions,
+} from './queueClient';
+
+export {
+  createGenerationWebSocketManager,
+  type GenerationWebSocketManager,
+  type GenerationWebSocketManagerOptions,
+} from './websocketManager';
+
+export {
+  type GenerationParamOverrides,
+  type GenerationRequestBody,
+  resolveGenerationBaseUrl,
+  resolveBackendUrl,
+  resolveGenerationRoute,
+  createGenerationParams,
+  requestGeneration,
+  fetchActiveGenerationJobs,
+  toGenerationRequestPayload,
+  startGeneration,
+  cancelGenerationJob,
+  deleteGenerationResult,
+  downloadGenerationResult,
+} from './generationService';
+
+export {
+  ensureArray,
+  logValidationIssues,
+  parseGenerationJobStatuses,
+  parseGenerationResults,
+} from './validation';
